[REFACTOR] Phase 5: one splitFunctionArguments (task 5.3)
Extract the bracket-depth-aware implementation (previously a private
copy in actionBranch.ts, a strict superset of datetime.ts's exported
quote-only version) into utils/functionArgs.ts and use it from both.
Removes the same-name/different-semantics trap. Added unit tests
covering the datetime call-site shapes (quoted patterns with commas,
escapes, nesting, blanks) - all pass against the shared version.
(The Swift interpreter mirror stays by design.)

Verified: web build, lint, 160/160 unit tests.
